test: move unit tests to a new folder
