+++ packages/contracts/test/unit/curation/configuration.test.ts
