+++ packages/contracts/test/unit/lib/gnsUtils.ts
